+++ types/account_test.go
@@ -1,11 +1,11 @@
-package types
+package types_test
 
 import (
 	"encoding/json"
 	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 
 	tmamino "github.com/tendermint/tendermint/crypto/encoding/amino"
 	"github.com/tendermint/tendermint/crypto/secp256k1"
@@ -13,65 +13,101 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 
-	emintcrypto "github.com/cosmos/ethermint/crypto"
+	"github.com/cosmos/ethermint/crypto"
+	"github.com/cosmos/ethermint/types"
 )
 
 func init() {
-	tmamino.RegisterKeyType(emintcrypto.PubKeySecp256k1{}, emintcrypto.PubKeyAminoName)
-	tmamino.RegisterKeyType(emintcrypto.PrivKeySecp256k1{}, emintcrypto.PrivKeyAminoName)
+	tmamino.RegisterKeyType(crypto.PubKeySecp256k1{}, crypto.PubKeyAminoName)
+	tmamino.RegisterKeyType(crypto.PrivKeySecp256k1{}, crypto.PrivKeyAminoName)
 }
 
-func TestEthermintAccountJSON(t *testing.T) {
+type AccountTestSuite struct {
+	suite.Suite
+
+	account *types.EthAccount
+}
+
+func (suite *AccountTestSuite) SetupTest() {
 	pubkey := secp256k1.GenPrivKey().PubKey()
 	addr := sdk.AccAddress(pubkey.Address())
-	balance := sdk.NewCoins(NewPhotonCoin(sdk.OneInt()))
+	balance := sdk.NewCoins(types.NewPhotonCoin(sdk.OneInt()))
 	baseAcc := auth.NewBaseAccount(addr, balance, pubkey, 10, 50)
-	ethAcc := EthAccount{BaseAccount: baseAcc, CodeHash: []byte{1, 2}}
-
-	bz, err := json.Marshal(ethAcc)
-	require.NoError(t, err)
-
-	bz1, err := ethAcc.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, string(bz1), string(bz))
-
-	var a EthAccount
-	require.NoError(t, a.UnmarshalJSON(bz))
-	require.Equal(t, ethAcc.String(), a.String())
-	require.Equal(t, ethAcc.PubKey, a.PubKey)
+	suite.account = &types.EthAccount{
+		BaseAccount: baseAcc,
+		CodeHash:    []byte{1, 2},
+	}
 }
 
-func TestEthermintPubKeyJSON(t *testing.T) {
-	privkey, err := emintcrypto.GenerateKey()
-	require.NoError(t, err)
+func TestAccountTestSuite(t *testing.T) {
+	suite.Run(t, new(AccountTestSuite))
+}
+
+func (suite *AccountTestSuite) TestEthAccount_Balance() {
+
+	testCases := []struct {
+		name         string
+		denom        string
+		initialCoins sdk.Coins
+		amount       sdk.Int
+	}{
+		{"positive diff", types.AttoPhoton, sdk.Coins{}, sdk.OneInt()},
+		{"zero diff, same coin", types.AttoPhoton, sdk.NewCoins(types.NewPhotonCoin(sdk.ZeroInt())), sdk.ZeroInt()},
+		{"zero diff, other coin", sdk.DefaultBondDenom, sdk.NewCoins(types.NewPhotonCoin(sdk.ZeroInt())), sdk.ZeroInt()},
+		{"negative diff", types.AttoPhoton, sdk.NewCoins(types.NewPhotonCoin(sdk.NewInt(10))), sdk.NewInt(1)},
+	}
+
+	for _, tc := range testCases {
+		suite.Run(tc.name, func() {
+			suite.SetupTest() // reset values
+			suite.account.SetCoins(tc.initialCoins)
+
+			suite.account.SetBalance(tc.denom, tc.amount)
+			suite.Require().Equal(tc.amount, suite.account.Balance(tc.denom))
+		})
+	}
+
+}
+
+func (suite *AccountTestSuite) TestEthermintAccountJSON() {
+	bz, err := json.Marshal(suite.account)
+	suite.Require().NoError(err)
+
+	bz1, err := suite.account.MarshalJSON()
+	suite.Require().NoError(err)
+	suite.Require().Equal(string(bz1), string(bz))
+
+	var a types.EthAccount
+	suite.Require().NoError(a.UnmarshalJSON(bz))
+	suite.Require().Equal(suite.account.String(), a.String())
+	suite.Require().Equal(suite.account.PubKey, a.PubKey)
+}
+
+func (suite *AccountTestSuite) TestEthermintPubKeyJSON() {
+	privkey, err := crypto.GenerateKey()
+	suite.Require().NoError(err)
 	bz := privkey.PubKey().Bytes()
 
 	pubk, err := tmamino.PubKeyFromBytes(bz)
-	require.NoError(t, err)
-	require.Equal(t, pubk, privkey.PubKey())
+	suite.Require().NoError(err)
+	suite.Require().Equal(pubk, privkey.PubKey())
 }
 
-func TestSecpPubKeyJSON(t *testing.T) {
+func (suite *AccountTestSuite) TestSecpPubKeyJSON() {
 	pubkey := secp256k1.GenPrivKey().PubKey()
 	bz := pubkey.Bytes()
 
 	pubk, err := tmamino.PubKeyFromBytes(bz)
-	require.NoError(t, err)
-	require.Equal(t, pubk, pubkey)
+	suite.Require().NoError(err)
+	suite.Require().Equal(pubk, pubkey)
 }
 
-func TestEthermintAccount_String(t *testing.T) {
-	pubkey := secp256k1.GenPrivKey().PubKey()
-	addr := sdk.AccAddress(pubkey.Address())
-	balance := sdk.NewCoins(NewPhotonCoin(sdk.OneInt()))
-	baseAcc := auth.NewBaseAccount(addr, balance, pubkey, 10, 50)
-	ethAcc := EthAccount{BaseAccount: baseAcc, CodeHash: []byte{1, 2}}
-
+func (suite *AccountTestSuite) TestEthermintAccount_String() {
 	config := sdk.GetConfig()
-	SetBech32Prefixes(config)
+	types.SetBech32Prefixes(config)
 
-	bech32pubkey, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, pubkey)
-	require.NoError(t, err)
+	bech32pubkey, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, suite.account.PubKey)
+	suite.Require().NoError(err)
 
 	accountStr := fmt.Sprintf(`|
   address: %s
@@ -83,66 +119,60 @@ func TestEthermintAccount_String(t *testing.T) {
   account_number: 10
   sequence: 50
   code_hash: "0102"
-`, addr, ethAcc.EthAddress().String(), bech32pubkey)
+`, suite.account.Address, suite.account.EthAddress().String(), bech32pubkey)
 
-	require.Equal(t, accountStr, ethAcc.String())
+	suite.Require().Equal(accountStr, suite.account.String())
 
-	i, err := ethAcc.MarshalYAML()
-	require.NoError(t, err)
+	i, err := suite.account.MarshalYAML()
+	suite.Require().NoError(err)
 
 	var ok bool
 	accountStr, ok = i.(string)
-	require.True(t, ok)
-	require.Contains(t, accountStr, addr.String())
-	require.Contains(t, accountStr, bech32pubkey)
+	suite.Require().True(ok)
+	suite.Require().Contains(accountStr, suite.account.Address.String())
+	suite.Require().Contains(accountStr, bech32pubkey)
 }
 
-func TestEthermintAccount_MarshalJSON(t *testing.T) {
-	pubkey := secp256k1.GenPrivKey().PubKey()
-	addr := sdk.AccAddress(pubkey.Address())
-	balance := sdk.NewCoins(NewPhotonCoin(sdk.OneInt()))
-	baseAcc := auth.NewBaseAccount(addr, balance, pubkey, 10, 50)
-	ethAcc := &EthAccount{BaseAccount: baseAcc, CodeHash: []byte{1, 2}}
+func (suite *AccountTestSuite) TestEthermintAccount_MarshalJSON() {
+	bz, err := suite.account.MarshalJSON()
+	suite.Require().NoError(err)
+	suite.Require().Contains(string(bz), suite.account.EthAddress().String())
 
-	bz, err := ethAcc.MarshalJSON()
-	require.NoError(t, err)
-	require.Contains(t, string(bz), ethAcc.EthAddress().String())
-
-	res := new(EthAccount)
+	res := new(types.EthAccount)
 	err = res.UnmarshalJSON(bz)
-	require.NoError(t, err)
-	require.Equal(t, ethAcc, res)
+	suite.Require().NoError(err)
+	suite.Require().Equal(suite.account, res)
 
-	bech32pubkey, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, pubkey)
-	require.NoError(t, err)
+	bech32pubkey, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeAccPub, suite.account.PubKey)
+	suite.Require().NoError(err)
 
 	// test that the sdk.AccAddress is populated from the hex address
 	jsonAcc := fmt.Sprintf(
 		`{"address":"","eth_address":"%s","coins":[{"denom":"aphoton","amount":"1"}],"public_key":"%s","account_number":10,"sequence":50,"code_hash":"0102"}`,
-		ethAcc.EthAddress().String(), bech32pubkey,
+		suite.account.EthAddress().String(), bech32pubkey,
 	)
 
-	res = new(EthAccount)
+	res = new(types.EthAccount)
 	err = res.UnmarshalJSON([]byte(jsonAcc))
-	require.NoError(t, err)
-	require.Equal(t, addr.String(), res.Address.String())
+	suite.Require().NoError(err)
+	suite.Require().Equal(suite.account.Address.String(), res.Address.String())
 
 	jsonAcc = fmt.Sprintf(
 		`{"address":"","eth_address":"","coins":[{"denom":"aphoton","amount":"1"}],"public_key":"%s","account_number":10,"sequence":50,"code_hash":"0102"}`,
 		bech32pubkey,
 	)
 
-	res = new(EthAccount)
+	res = new(types.EthAccount)
 	err = res.UnmarshalJSON([]byte(jsonAcc))
-	require.Error(t, err, "should fail if both address are empty")
+	suite.Require().Error(err, "should fail if both address are empty")
 
 	// test that the sdk.AccAddress is populated from the hex address
 	jsonAcc = fmt.Sprintf(
 		`{"address": "%s","eth_address":"0x0000000000000000000000000000000000000000","coins":[{"denom":"aphoton","amount":"1"}],"public_key":"%s","account_number":10,"sequence":50,"code_hash":"0102"}`,
-		ethAcc.Address.String(), bech32pubkey,
+		suite.account.Address.String(), bech32pubkey,
 	)
 
-	res = new(EthAccount)
+	res = new(types.EthAccount)
 	err = res.UnmarshalJSON([]byte(jsonAcc))
-	require.Error(t, err, "should fail if addresses mismatch")
+	suite.Require().Error(err, "should fail if addresses mismatch")
 }