+++ types/config_test.go
@@ -10,7 +10,6 @@ import (
 
 func TestSetBech32Prefixes(t *testing.T) {
 	config := sdk.GetConfig()
-	config = sdk.NewConfig() // reset config values
 
 	require.Equal(t, sdk.Bech32PrefixAccAddr, config.GetBech32AccountAddrPrefix())
 	require.Equal(t, sdk.Bech32PrefixAccPub, config.GetBech32AccountPubPrefix())