+++ x/evm/types/statedb.go
@@ -693,7 +693,8 @@ func (csdb *CommitStateDB) Prepare(thash, bhash ethcmn.Hash, txi int) {
 func (csdb *CommitStateDB) CreateAccount(addr ethcmn.Address) {
 	newobj, prevobj := csdb.createObject(addr)
 	if prevobj != nil {
-		newobj.setBalance(sdk.NewIntFromBigInt(prevobj.Balance()))
+		evmDenom := csdb.GetParams().EvmDenom
+		newobj.setBalance(evmDenom, sdk.NewIntFromBigInt(prevobj.Balance()))
 	}
 }